requirements.txt
absl-py==0.6.1
astor==0.7.1
cachetools==3.0.0
gast==0.2.0
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
grpcio==1.17.0
h5py==2.8.0
httplib2==0.12.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mpmath==1.1.0
numpy==1.15.4
oauth2client==4.1.3
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyYAML==3.13
rsa==4.0
scipy==1.1.0
six==1.12.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
Werkzeug==0.14.1
